Rename einsum_impl.h (#73801)
