+++ paddle/phi/kernels/gpu/einsum_grad_kernel.cu
@@ -15,7 +15,7 @@
 #include "paddle/phi/backends/gpu/gpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/kernels/einsum_kernel.h"
-#include "paddle/phi/kernels/impl/einsum_grad_impl.h"
+#include "paddle/phi/kernels/impl/einsum_grad_kernel_impl.h"
 
 PD_REGISTER_KERNEL(einsum_grad,
                    GPU,
