+++ paddle/phi/kernels/gpu/einsum_kernel.cu
@@ -16,7 +16,7 @@
 
 #include "paddle/phi/backends/gpu/gpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
-#include "paddle/phi/kernels/impl/einsum_impl.h"
+#include "paddle/phi/kernels/impl/einsum_kernel_impl.h"
 
 PD_REGISTER_KERNEL(einsum,
                    GPU,
