+++ paddle/phi/kernels/xpu/einsum_grad_kernel.cc
@@ -15,7 +15,7 @@
 #include "paddle/phi/kernels/einsum_grad_kernel.h"
 
 #include "paddle/phi/core/kernel_registry.h"
-#include "paddle/phi/kernels/impl/einsum_grad_impl.h"
+#include "paddle/phi/kernels/impl/einsum_grad_kernel_impl.h"
 
 PD_REGISTER_KERNEL(einsum_grad, XPU, ALL_LAYOUT, phi::EinsumGradKernel, float) {
 }
