+++ paddle/phi/kernels/xpu/einsum_kernel.cc
@@ -17,7 +17,7 @@
 #include "paddle/phi/backends/xpu/enforce_xpu.h"
 #include "paddle/phi/core/kernel_registry.h"
 
-#include "paddle/phi/kernels/impl/einsum_impl.h"
+#include "paddle/phi/kernels/impl/einsum_kernel_impl.h"
 
 PD_REGISTER_KERNEL(
     einsum_infer, XPU, ALL_LAYOUT, phi::EinsumInferKernel, float) {}
